requirements.txt
GEOparse~=2.0.2
pandas~=1.1.4
numpy~=1.19.3
seaborn~=0.11.0
matplotlib~=3.3.3
scikit-learn~=0.24.0
umap-learn~=0.5.1
cmapPy~=4.0.1
xlrd~=1.2.0
joblib~=0.17.0